Extract ChangePayload of HistoryChangePack
